refactor: CreditCard 도메인 로직으로 이동
- 마스킹과 같은 비즈니스 로직의 존재로 이동
- 배열보다 페이먼츠의 타입을 사용하게 함
