+++ src/pages/payments/types.ts
@@ -1,3 +1,10 @@
+export type CardNumbers = {
+  first: string;
+  second: string;
+  third: string;
+  fourth: string;
+};
+
 export type ExpirationDate = {
   year: string;
   month: string;
